# frontend/openapi/contrato-provisorio.yaml
...
openapi: 3.0.3
info:
  title: SIAFIC — Execução orçamentária/financeira (derivado do código real)
  version: 0.3.0
  description: >
    Empenho/liquidação/pagamento (escrita, RAZ-105) + execução orçamentária agregada
    (consulta, RAZ-101) + assinatura eletrônica gov.br do empenho (RAZ-141/ADR-0039).
    enteId sempre no path (ADR-0015). Autenticação: Authorization: Bearer <asserção gov.br>,
    validada a cada request (stateless) por ServicoIdentidade — exceto os 3 endpoints de
    assinatura, que usam HttpSession/cookie (ADR-0017, única exceção documentada ao §6.1).
servers:
  - url: /api/v1/entes/{enteId}
    variables:
      enteId:
        default: '00000000-0000-0000-0000-000000000000'
security:
  - govbrBearer: []
paths:
...
  /execucao/liquidacoes:
    get:
      operationId: consultarFilaAprovacao
      summary: >-
        Fila de aprovação de liquidações por statusAprovacao, cursor opaco, PII do credor
        não exposta (ADR-0029 §1). # java: consulta.ExecucaoConsultaController.fila
      parameters:
        - name: statusAprovacao
          in: query
          required: false
          schema: { type: string, enum: [pendente, aprovada, devolvida], default: pendente }
        - name: cursor
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer }
        - name: fonte
          in: query
          required: false
          schema: { type: string }
        - name: dataInicio
          in: query
          required: false
          schema: { type: string, format: date }
        - name: dataFim
          in: query
          required: false
          schema: { type: string, format: date }
        - name: valorMin
          in: query
          required: false
          schema: { type: string }
        - name: valorMax
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Página da fila de aprovação.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FilaAprovacaoResponse' }
        '401': { $ref: '#/components/responses/ErroContrato' }
        '403': { $ref: '#/components/responses/ErroContrato' }
    post:
      operationId: registrarLiquidacao
      summary: Registra uma liquidação vinculada a um empenho. # java: escrita.LiquidacaoController.registrar
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LiquidacaoRequest' }
      responses:
        '201':
          description: Liquidação registrada.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LiquidacaoResponse' }
        '400': { $ref: '#/components/responses/ErroContrato' }
        '401': { $ref: '#/components/responses/ErroContrato' }
        '409': { $ref: '#/components/responses/ErroContrato' }
...
  /execucao/pagamentos:
    post:
      operationId: registrarPagamento
      summary: Registra um pagamento vinculado a uma liquidação. # java: escrita.PagamentoController.registrar
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PagamentoRequest' }
      responses:
        '201':
          description: Pagamento registrado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagamentoResponse' }
        '400': { $ref: '#/components/responses/ErroContrato' }
        '401': { $ref: '#/components/responses/ErroContrato' }
        '409': { $ref: '#/components/responses/ErroContrato' }
...
components:
  securitySchemes:
    govbrBearer:
      type: http
      scheme: bearer
      description: >
        java: consulta.SessaoAutenticadaHttpResolver — "Authorization: Bearer <asserção gov.br>",
        verificada a cada request via ServicoIdentidade.autenticar (stateless, sem cookie/sessão).
    assinaturaSessao:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: >
        java: assinatura.AssinaturaGovBrOAuthController / assinatura.AssinaturaEmpenhoController —
        HttpSession/cookie (RepositorioSessaoAssinaturaGovBr guarda state/PKCE/token na sessão),
        NÃO Authorization: Bearer. Única exceção documentada ao security scheme padrão deste
        contrato (§6.1) — ADR-0017, decisão 5 do ADR-0039.
  responses:
    ErroContrato:
      description: Envelope único de erro (RAZ-79 §6.1 / ADR-0030 §3). # java: consulta.ErroContratoExceptionHandler
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErroContratoResponse' }
  schemas:
    Dinheiro:
      type: string
      description: >
        Decimal (escala 2) serializado como string pelo DinheiroJacksonModule real
        (ADR-0030 §2) — nunca number no cliente.
      example: '128450.00'
    ErroContratoResponse:
      type: object
      required: [codigo, mensagem, detalhes]
      properties:
        codigo: { type: string, example: nao_autenticado }
        mensagem: { type: string }
        detalhes: { type: object, additionalProperties: { type: string } }
...
    LiquidacaoRequest:
      type: object
      required: [empenhoId, dataCompetencia, valor, documentosSuporte, historico]
      properties:
        empenhoId: { type: string, format: uuid }
        dataCompetencia: { type: string, format: date }
        valor: { $ref: '#/components/schemas/Dinheiro' }
        documentosSuporte:
          type: array
          items: { $ref: '#/components/schemas/DocumentoSuporte' }
        historico: { type: string }
    DocumentoSuporte:
      type: object
      required: [tipo, numero, dataEmissao]
      properties:
        tipo: { type: string }
        numero: { type: string }
        dataEmissao: { type: string, format: date }
        referenciaExterna: { type: string, nullable: true }
    LiquidacaoResponse:
      type: object
      required: [id, empenhoId, dataCompetencia, valor, documentosSuporte, historico, fatoContabilId, status]
      properties:
        id: { type: string, format: uuid }
        empenhoId: { type: string, format: uuid }
        dataCompetencia: { type: string, format: date }
        valor: { $ref: '#/components/schemas/Dinheiro' }
        documentosSuporte:
          type: array
          items: { $ref: '#/components/schemas/DocumentoSuporte' }
        historico: { type: string }
        fatoContabilId: { type: string, format: uuid }
        status: { type: string, enum: [pendente, aprovada, devolvida] }
    PagamentoRequest:
      type: object
      required: [liquidacaoId, dataCompetencia, valor, natureza, historico]
      properties:
        liquidacaoId: { type: string, format: uuid }
        dataCompetencia: { type: string, format: date }
        valor: { $ref: '#/components/schemas/Dinheiro' }
        natureza: { type: string }
        beneficiario: { $ref: '#/components/schemas/Beneficiario', nullable: true }
        ordemBancaria: { type: string, nullable: true }
        historico: { type: string }
    Beneficiario:
      type: object
      required: [nome, cpfCnpj]
      properties:
        nome: { type: string }
        cpfCnpj: { type: string, description: 'Mascarado na resposta — nunca em claro (RAZ-79 §6.1).' }
    PagamentoResponse:
      type: object
      required: [id, liquidacaoId, dataCompetencia, valor, natureza, historico, fatoContabilId]
      properties:
        id: { type: string, format: uuid }
        liquidacaoId: { type: string, format: uuid }
        dataCompetencia: { type: string, format: date }
        valor: { $ref: '#/components/schemas/Dinheiro' }
        natureza: { type: string }
        beneficiario: { $ref: '#/components/schemas/Beneficiario', nullable: true }
        ordemBancaria: { type: string, nullable: true }
        historico: { type: string }
        fatoContabilId: { type: string, format: uuid }
...
    ItemFilaAprovacao:
      type: object
      # java: consulta.ExecucaoConsultaController.ItemFilaResponse
      required:
        [id, empenhoId, numeroEmpenho, exercicioEmpenho, credorId, valor, dataCompetencia, statusAprovacao]
      properties:
        id: { type: string, format: uuid }
        empenhoId: { type: string, format: uuid }
        numeroEmpenho: { type: integer, format: int64 }
        exercicioEmpenho: { type: integer }
        credorId: { type: string, format: uuid }
        valor: { $ref: '#/components/schemas/Dinheiro' }
        dataCompetencia: { type: string, format: date }
        statusAprovacao: { type: string, enum: [pendente, aprovada, devolvida] }
    FilaAprovacaoResponse:
      type: object
      # java: consulta.ExecucaoConsultaController.FilaAprovacaoResponse
      required: [itens, proximoCursor]
      properties:
        itens:
          type: array
          items: { $ref: '#/components/schemas/ItemFilaAprovacao' }
        proximoCursor: { type: string, nullable: true }
...